Move 'core handlers' subcommand to 'handlers list-builtins'
