+++ commands/core/core.go
@@ -16,7 +16,6 @@ package core
 
 import (
 	"github.com/sacloud/autoscaler/commands/core/example"
-	"github.com/sacloud/autoscaler/commands/core/handlers"
 	"github.com/sacloud/autoscaler/commands/core/resources"
 	"github.com/sacloud/autoscaler/commands/core/start"
 	"github.com/sacloud/autoscaler/commands/core/validate"
@@ -29,7 +28,7 @@ var Command = &cobra.Command{
 	Short:         "A set of sub commands to manage autoscaler's core server",
 	SilenceErrors: true,
 
-	// Note: 互換性維持用(since v0.5)
+	// Memo: 互換性維持用(since v0.5) v1.0リリース時に除去する
 	// coreのサブコマンド群はroot直下に配置するが、互換性維持のためcoreサブコマンド配下にも配置する。
 	// このときにcoreサブコマンド自体を非表示にするようにHiddenを設定しておく
 	Hidden: true,
@@ -38,7 +37,6 @@ var Command = &cobra.Command{
 var subCommands = []*cobra.Command{
 	example.Command,
 	start.Command,
-	handlers.Command,
 	validate.Command,
 	resources.Command,
 }
